protocol: refactor conversion helpers
