+++ protocol/src/error.rs
@@ -1,0 +1,10 @@
+#[derive(Debug)]
+pub struct DecodeError;
+
+impl std::fmt::Display for DecodeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "failed to decode scalar")
+    }
+}
+
+impl error_stack::Context for DecodeError {}
