+++ protocol/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod client;
 pub mod dna;
+pub mod error;
 pub mod evm;
+mod helpers;
 pub mod starknet;
